Fix database types check
On #34 I made changes to database types without touching anything regarding supabase. This is odd enough, but what I found even stranger is that there were no complaints in CI.

This made me think that the CI job isn't actually doing what it is supposed to.

As an attempt to debug and potentially fix an issue with the CI job, I'm following this guide to re-create the task from scratch:
https://supabase.com/docs/guides/cli/github-action/generating-types